frontend_streamlit/pyproject.toml
[project]
name = "frontend"
version = "0.1.0"
description = "Streamlit frontend for dream team project"
readme = "README.md"
requires-python = ">=3.10, <3.13"
dependencies = [
    "streamlit>=1.28.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "plotly>=5.17.0",
    "streamlit-chat>=0.1.1",
    "streamlit-option-menu>=0.3.6",
    "azure-identity>=1.15.0",
    "msal-streamlit>=0.1.0"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"